# XW-2429 | Skip costly and unnecessary operations on file pages