@WikiaTech 2016-12-12
SonarQube analysis reported 10 issues

Note: The following issues were found on lines that were not modified in the pull request. Because these issues can't be reported as line comments, they are summarized here:

1. ![MAJOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-major.png) [MercuryApiController.class.php#L360](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L360): Reduce this "switch/case" number of lines from 19 to at most 10, for example by extracting code into function. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1151)
1. ![MAJOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-major.png) [MercuryApiController.class.php#L364](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L364): Refactor this code to not nest more than 4 "if", "for", "while", "switch" and "try" statements. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS134)
1. ![MAJOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-major.png) [MercuryApiController.class.php#L379](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L379): Reduce this "switch/case" number of lines from 16 to at most 10, for example by extracting code into function. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1151)
1. ![MAJOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-major.png) [MercuryApiController.class.php#L381](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L381): Refactor this code to not nest more than 4 "if", "for", "while", "switch" and "try" statements. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS134)
1. ![MAJOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-major.png) [MercuryApiController.class.php#L521](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L521): Remove this commented out code. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS125)
1. ![MINOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-minor.png) [MercuryApiController.class.php#L10](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L10): Define a constant instead of duplicating this literal "title" 3 times. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1192)
1. ![MINOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-minor.png) [MercuryApiController.class.php#L316](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L316): Define a constant instead of duplicating this literal "article" 5 times. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1192)
1. ![MINOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-minor.png) [MercuryApiController.class.php#L353](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L353): Define a constant instead of duplicating this literal "details" 6 times. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1192)
1. ![MINOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-minor.png) [MercuryApiController.class.php#L358](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L358): Replace this "switch" statement with "if" statements to increase readability. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1301)
1. ![MINOR](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/severity-minor.png) [MercuryApiController.class.php#L539](https://github.com/Wikia/app/blob/65dae8c03dfa070786574ad9239433eacad6c1cb/extensions/wikia/MercuryApi/MercuryApiController.class.php#L539): Define a constant instead of duplicating this literal "languageName" 3 times. [![rule](https://raw.githubusercontent.com/SonarCommunity/sonar-github/master/images/rule.png)](https://sonarqube.wikia-inc.com/coding_rules#rule_key=php%3AS1192)