@macbre 2016-12-13
And please note that SUS is actually The Solr Guy :) We want to get rid of tribal knowledge and document it well.