@rogatty 2016-12-13
Failing unit tests were caused by some network issue, they're fine after rebuilding the job.